Move cumulus indexing controller to base project
